Promote `AmLineChart` to be a smart component Dedicate `dumpChart` Add `watchStartChart` sagas Add `ChartReducer`
