+++ f1/src/sagas.js
@@ -4,14 +4,16 @@ import {watchLoginBtn} from "./frontpage/submit";
 import {watchCheckToken, watchGetUsername} from "./commons/sagas";
 import {watchSubmitBasic} from "./controlroom/sagas";
 import {watchSubmitAdvance} from "./controlroom/advance-sagas";
+import {watchStartChart} from "./displays/sagas/value-vs-time";
 
 export default function* rootSaga() {
-  yield all([
-    watchShootBacked(),
-    watchLoginBtn(),
-    watchGetUsername(),
-    watchCheckToken(),
-    watchSubmitBasic(),
-    watchSubmitAdvance(),
-  ])
+	yield all([
+		watchShootBacked(),
+		watchLoginBtn(),
+		watchGetUsername(),
+		watchCheckToken(),
+		watchSubmitBasic(),
+		watchSubmitAdvance(),
+		watchStartChart(),
+	])
 }
